+++ src/infrastructure/database/config.rs
@@ -20,7 +20,7 @@ impl DatabaseConfig {
                 item_id INTEGER PRIMARY KEY AUTOINCREMENT,
                 name TEXT NOT NULL,
                 description TEXT NOT NULL,
-                price INTEGER NOT NULL,
+                price REAL NOT NULL,
                 quantity INTEGER NOT NULL,
                 stock INTEGER NOT NULL,
                 category TEXT NOT NULL,
